added more documentation and changed one method

+++ main-omm/database/Code/editJsonFile.py
@@ -7,7 +7,7 @@ def findHighestFileNum(filePath, file): #this is specifically for question files
     newestFile = None
     searching = True
     
-    while searching:
+    while searching: #This while loop searches for the most recent iteration of a file by looking at the '-#' at the end of each file and returns the one with the highest one
         currentFile = str(file) + "-" + str(num) + ".json"
         if currentFile in os.listdir(filePath):
             newestFile = currentFile
@@ -16,8 +16,8 @@ def findHighestFileNum(filePath, file): #this is specifically for question files
             searching = False
     return newestFile
 
-def editAllJson():
-    filePath = Path('C:/Users/ponzo/OneDrive/Desktop/Code/questions/')
+def editAllJson(): #method for editing all files if needed, currently unused
+    filePath = Path('main-omm/database/Code/accounts')
     for file in os.listdir(filePath):
         fullPath = filePath / file
         with open(fullPath, 'r') as f:
@@ -28,7 +28,7 @@ def editAllJson():
             json.dump(data, f, indent=4)  # Save the updated data back to the file
 
 def editJsonAccountDetails(fileName, newEmail, newPassword, newFirstName, newLastName, newAccountType):
-    filePath = Path('C:/Users/ponzo/OneDrive/Desktop/Code/accounts/')
+    filePath = Path('main-omm/database/Code/accounts')
     fileName = filePath / f"{fileName}.json"
 
     if not os.path.exists(filePath / fileName):
@@ -52,7 +52,7 @@ def editJsonAccountDetails(fileName, newEmail, newPassword, newFirstName, newLas
             json.dump(data, f, indent=4)  #Save the updated data back to the file
 
 def editJsonAccountStatistics(fileName, newTotalQuestionsAnswered, newTotalQuestionsCorrect, newTotalTestsCompleted, newAverageScore):
-    filePath = Path('C:/Users/ponzo/OneDrive/Desktop/Code/accounts/')
+    filePath = Path('main-omm/database/Code/accounts')
     fileName = filePath / f"{fileName}.json"
 
     if not os.path.exists(filePath / fileName):
@@ -62,19 +62,19 @@ def editJsonAccountStatistics(fileName, newTotalQuestionsAnswered, newTotalQuest
         with open(filePath / fileName, 'r') as f:
             data = json.load(f)
             if newTotalQuestionsAnswered != None:
-                data['statistics']['totalQuestionsAnswered'] = newTotalQuestionsAnswered  # Edit the 'answerOneText' field
+                data['statistics']['totalQuestionsAnswered'] = newTotalQuestionsAnswered  #Set new amount of questions answered
             if newTotalQuestionsCorrect != None:
-                data['statistics']['totalQuestionsCorrect'] = newTotalQuestionsCorrect  # Edit the 'answerOneText' field
+                data['statistics']['totalQuestionsCorrect'] = newTotalQuestionsCorrect  #Set new amount of questions answered correctly
             if newTotalTestsCompleted != None:
-                data['statistics']['totalTestsCompleted'] = newTotalTestsCompleted  # Edit the 'answerOneText' field
+                data['statistics']['totalTestsCompleted'] = newTotalTestsCompleted  #Set new amount of total tests completed
             if newAverageScore != None:
-                data['statistics']['averageScore'] = newAverageScore  # Edit the 'answerOneText' field
+                data['statistics']['averageScore'] = newAverageScore  #Update average score
 
         with open(filePath / fileName, 'w') as f:
             json.dump(data, f, indent=4)  # Save the updated data back to the file
 
-def editJsonQuestion(fileName, newTitle, newQuestionText, newTags, newImagePath, allowImagePath, newAnswerOneText, newAnswerTwoText, newAnswerThreeText, allowAnswerThree, newAnswerFourText, allowAnswerFour, newAnswerFiveText, allowAnswerFive, newIsCorrect, newTotalTimesAnswered, newTotalTimesAnsweredCorrect, newTotalTimeSpentOnQuestion):
-    filePath = Path('C:/Users/ponzo/OneDrive/Desktop/Code/questions/')
+def editJsonQuestion(fileName, newTitle, newQuestionText, newTags, newImagePath, allowImagePath, newAnswerOneText, newAnswerTwoText, newAnswerThreeText, allowAnswerThree, newAnswerFourText, allowAnswerFour, newAnswerFiveText, allowAnswerFive, newIsCorrect):
+    filePath = Path('main-omm/database/Code/questions')
     baseFile = filePath / f"{fileName}-1.json"
 
     if not os.path.exists(baseFile): #make sure one file exists atleast
@@ -92,11 +92,11 @@ def editJsonQuestion(fileName, newTitle, newQuestionText, newTags, newImagePath,
                 data['questionDetails']['questionText'] = newQuestionText  #Sets new question text if needed
             if newTags != None:
                 data['questionDetails']['tags'] = newTags  #Sets new tags if needed
-            if data['questionDetails']['imagePath'] == None and allowImagePath:
+            if data['questionDetails']['imagePath'] == None and allowImagePath: #if image doesnt exist but you want to add one
                 data['questionDetails']['imagePath'] = newImagePath  
-            elif data['questionDetails']['imagePath'] != None and not allowImagePath:
+            elif data['questionDetails']['imagePath'] != None and not allowImagePath: #if image exists but want to remove it
                 data['questionDetails']['imagePath'] = None 
-            elif data['questionDetails']['imagePath'] != None and allowImagePath:
+            elif data['questionDetails']['imagePath'] != None and allowImagePath: # if image exists but want to update it
                 data['questionDetails']['imagePath'] = newImagePath
 
             data['questionDetails']['questionVersion'] += 1  #Question is being updated, so add 1 to question version
@@ -104,23 +104,23 @@ def editJsonQuestion(fileName, newTitle, newQuestionText, newTags, newImagePath,
             if newAnswerOneText != None:
                 data['answers']['answerOneText'] = newAnswerOneText  # Edit the 'answerOneText' field
             if newAnswerTwoText != None:
-                data['answers']['answerTwoText'] = newAnswerTwoText  # Edit the 'answerOneText' field
+                data['answers']['answerTwoText'] = newAnswerTwoText  # Edit the 'answerOneTwo' field
 
-            if data['answers']['answerThreeText'] == None and allowAnswerThree:
+            if data['answers']['answerThreeText'] == None and allowAnswerThree: #Same as image but instead for answer 3
                 data['answers']['answerThreeText'] = newAnswerThreeText  
             elif data['answers']['answerThreeText'] != None and not allowAnswerThree:
                 data['answers']['answerThreeText'] = None 
             elif data['answers']['answerThreeText'] != None and allowAnswerThree:
                 data['answers']['answerThreeText'] = newAnswerThreeText
 
-            if data['answers']['answerFourText'] == None and allowAnswerFour:
+            if data['answers']['answerFourText'] == None and allowAnswerFour: #Same as image but instead for answer 4
                 data['answers']['answerFourText'] = newAnswerFourText
             elif data['answers']['answerFourText'] != None and not allowAnswerFour:
                 data['answers']['answerFourText'] = None 
             elif data['answers']['answerFourText'] != None and allowAnswerFour:
                 data['answers']['answerFourText'] = newAnswerFourText
 
-            if data['answers']['answerFiveText'] == None and allowAnswerFive:
+            if data['answers']['answerFiveText'] == None and allowAnswerFive: #Same as image but instead for answer 5
                 data['answers']['answerFiveText'] = newAnswerFiveText 
             elif data['answers']['answerFiveText'] != None and not allowAnswerFive:
                 data['answers']['answerFiveText'] = None
@@ -128,17 +128,33 @@ def editJsonQuestion(fileName, newTitle, newQuestionText, newTags, newImagePath,
                 data['answers']['answerFiveText'] = newAnswerFiveText
 
             if newIsCorrect != None:
-                data['answers']['isCorrect'] = newIsCorrect  # Edit the 'answerOneText' field
-
-        
-            if newTotalTimesAnswered != None:
-                data['statistics']['totalTimesAnswered'] = newTotalTimesAnswered  # Edit the 'answerOneText' field
-            if newTotalTimesAnsweredCorrect != None:
-                data['statistics']['totalTimesAnsweredCorrect'] = newTotalTimesAnsweredCorrect  # Edit the 'answerOneText' field
-            if newTotalTimeSpentOnQuestion != None:
-                data['statistics']['totalTimeSpentOnQuestion'] = newTotalTimeSpentOnQuestion  # Edit the 'answerOneText' field
+                data['answers']['isCorrect'] = newIsCorrect  #Edit which answers are correct
 
             newFileName = fileName[:-6] + str(data['questionDetails']['questionVersion']) + ".json"
+
+        with open(filePath / newFileName, 'w') as f:
+            json.dump(data, f, indent=4)  # Save the updated data back to the file
+
+def editJsonQuestionStatistics(newTotalTimesAnswered, newTotalTimesAnsweredCorrect, newTotalTimeSpentOnQuestion):
+
+    filePath = Path('main-omm/database/Code/questions')
+    baseFile = filePath / f"{fileName}-1.json"
+
+    if not os.path.exists(baseFile): #make sure one file exists atleast
+        print("File does not exist")
+
+    else:
+        fileName = findHighestFileNum(filePath, fileName) #find most recent version
+        print(fileName + " is most recent")
+        with open(filePath / fileName, 'r') as f:
+            data = json.load(f)
+        
+            if newTotalTimesAnswered != None:
+                data['statistics']['totalTimesAnswered'] = newTotalTimesAnswered  #Update total times question has been answered
+            if newTotalTimesAnsweredCorrect != None:
+                data['statistics']['totalTimesAnsweredCorrect'] = newTotalTimesAnsweredCorrect  #Update total times question has been correctly answered
+            if newTotalTimeSpentOnQuestion != None:
+                data['statistics']['totalTimeSpentOnQuestion'] = newTotalTimeSpentOnQuestion  #Update total time spent on question
 
         with open(filePath / newFileName, 'w') as f:
             json.dump(data, f, indent=4)  # Save the updated data back to the file
